+++ HISTORY.rst
@@ -2,6 +2,8 @@
 History
 =======
 
+* Add fixer for removal of compatibility imports for ``django.core.exceptions.EmptyResultSet``.
+
 1.2.0 (2021-09-02)
 ------------------